+++ app/controllers/open_api/v1/invoices_controller.rb
@@ -21,7 +21,8 @@ class OpenAPI::V1::InvoicesController < OpenAPI::V1::BaseController
   end
 
   private
-    def per_page
-      params[:per_page] || 20
-    end
+
+  def per_page
+    params[:per_page] || 20
+  end
 end