+++ app/services/reservations/reserve.rb
@@ -13,24 +13,28 @@ class Reservations::Reserve
   # Confirm the payment of the given reservation, generate the associated documents and save teh record into
   # the database.
   ##
-  def pay_and_save(reservation, payment_details: nil, payment_intent_id: nil, schedule: false, payment_method: nil)
+  def pay_and_save(reservation, payment_details: nil, intent_id: nil, schedule: false, payment_method: nil)
     user = User.find(user_id)
     reservation.statistic_profile_id = StatisticProfile.find_by(user_id: user_id).id
 
-    reservation.pre_check
-    payment = if schedule
-                generate_schedule(reservation: reservation,
-                                  total: payment_details[:before_coupon],
-                                  operator_profile_id: operator_profile_id,
-                                  user: user,
-                                  payment_method: payment_method,
-                                  coupon_code: payment_details[:coupon])
-              else
-                generate_invoice(reservation, operator_profile_id, payment_details, payment_intent_id)
-              end
-    payment.save
-    WalletService.debit_user_wallet(payment, user, reservation)
-    reservation.post_save
+    ActiveRecord::Base.transaction do
+      reservation.pre_check
+      payment = if schedule
+                  generate_schedule(reservation: reservation,
+                                    total: payment_details[:before_coupon],
+                                    operator_profile_id: operator_profile_id,
+                                    user: user,
+                                    payment_method: payment_method,
+                                    coupon_code: payment_details[:coupon],
+                                    setup_intent_id: intent_id)
+                else
+                  generate_invoice(reservation, operator_profile_id, payment_details, intent_id)
+                end
+      payment.save
+      reservation.save
+      WalletService.debit_user_wallet(payment, user, reservation)
+      reservation.post_save
+    end
     true
   end
 
@@ -39,7 +43,8 @@ class Reservations::Reserve
   ##
   # Generate the invoice for the given reservation+subscription
   ##
-  def generate_schedule(reservation: nil, total: nil, operator_profile_id: nil, user: nil, payment_method: nil, coupon_code: nil)
+  def generate_schedule(reservation: nil, total: nil, operator_profile_id: nil, user: nil, payment_method: nil, coupon_code: nil,
+                        setup_intent_id: nil)
     operator = InvoicingProfile.find(operator_profile_id)&.user
     coupon = Coupon.find_by(code: coupon_code) unless coupon_code.nil?
 
@@ -50,7 +55,8 @@ class Reservations::Reserve
       operator: operator,
       payment_method: payment_method,
       user: user,
-      reservation: reservation
+      reservation: reservation,
+      setup_intent_id: setup_intent_id
     )
   end